test(automations): split automations-specific strategies into separate module
Separates generic MongoDB filter strategies from automations-specific ones
so the generic module can later move out of test_automations as a clean
file rename.
