+++ tests/unit_tests/test_automations/test_run_events.py
@@ -20,10 +20,9 @@ from wandb.automations._filters.run_metrics import Agg, ChangeDir, MetricAgg, Me
 from wandb.automations._filters.run_states import ReportedRunState
 from wandb.automations.events import StateFilter
 
-from ._strategies import (
+from ._automations_strategies import (
     aggs,
     cmp_keys,
-    ints_or_floats,
     metric_change_filters,
     metric_names,
     metric_zscore_filters,
@@ -33,6 +32,7 @@ from ._strategies import (
     run_states,
     window_sizes,
 )
+from ._strategies import ints_or_floats
 
 
 @composite
